config: Store Ansible Galaxy roles outside
Store the roles outside the repo by default. This was changed in f78132b6 but did not include a description of why this change was made. 

The path was then duplicated which leads me to think it was maybe not intentional. If we are going to start vendoring of roles we should call that out explicitly.